requirements_dev.txt
flake8==2.4.1
tox==2.1.1
tox-pyenv==1.0.3
coverage==4.0
flask_sqlalchemy
pytest
-e .
